complete 3rd draft
I await testing with the 3rd form of this. But I'm feeling really good
about it. It stacks hexagonally, and has some hands poking out the
bottom that grab the cradle below it.

I also built a foot, to bootstrap rising these things out of the ground.

It is 100% printable, but it will require gluing. Most anyone who has
these paints should also have glue...