make a smaller cradle
The cradle size makes it hard to grasp the bottles. I need smaller
cradles and this appears to do it.

There is a problem with the radius, apothem, or both. My radial
calculations are just a little bit off and it makes some transformations
slightly incorrect.

There's cleanup done here, but more cleanup needed.