hold the bottle, but leave enough space to grasp
I found the paint bottle was difficult to pull out of the cradle,
especially once the cradle was amongst an upper layer of cradles. This
should leave plenty of bottle poking out. My only concern is that it
might not have a great center of gravity anymore. I need to test this.